# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Room Reservation",
        "description": "API Documentation for Hotel Room Reservation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reservations": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Tampilkan semua reservasi",
                "description": "Admin melihat semua reservasi. User hanya melihat reservasi miliknya sendiri.",
                "operationId": "f10e68ff0ce263ef754b2f640eda8d93",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data reservasi"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Buat reservasi baru",
                "description": "User membuat reservasi untuk dirinya sendiri. Admin bisa membuat reservasi untuk user lain dengan menyertakan 'user_id'.",
                "operationId": "4f59911f11563f79b4f01bd02222607d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "room_id",
                                    "check_in_date",
                                    "check_out_date"
                                ],
                                "properties": {
                                    "room_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "check_in_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-20"
                                    },
                                    "check_out_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-22"
                                    },
                                    "extra_beds": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "description": "[Khusus Admin] ID user yang melakukan reservasi.",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservasi berhasil dibuat"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Tampilkan detail reservasi",
                "description": "Admin dapat melihat detail reservasi manapun. User hanya dapat melihat miliknya.",
                "operationId": "14c077250033bb10a7305f9724f10bd4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Reservasi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail reservasi"
                    },
                    "403": {
                        "description": "Forbidden (tidak punya hak akses)"
                    },
                    "404": {
                        "description": "Not Found (data tidak ditemukan)"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Update reservasi",
                "description": "Admin bisa update semua field. User hanya bisa update check_in_date, check_out_date, dan extra_beds.",
                "operationId": "453c2db0a909455ff1a629a0ed80f9a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Reservasi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "check_in_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-21"
                                    },
                                    "extra_beds": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "status": {
                                        "description": "[Khusus Admin] Status reservasi.",
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "confirmed",
                                            "cancelled",
                                            "CheckIn",
                                            "CheckOut"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reservasi berhasil diupdate"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Hapus reservasi",
                "description": "Menghapus reservasi. Hanya admin atau user yang memiliki reservasi yang bisa menghapus.",
                "operationId": "3f3a1d6e7ffee88b7ef58128a7f9295c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Reservasi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservasi berhasil dihapus"
                    },
                    "403": {
                        "description": "Forbidden (bukan pemilik atau admin)"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}